requirements.txt
blinker==1.7.0
Brotli==1.1.0
click==8.1.7
colorama==0.4.6
flask==3.0.2
Flask-Compress==1.14
Flask-Cors==4.0.0
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.24.4
pandas==2.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
tzdata==2024.1
werkzeug==3.0.1
zipp==3.17.0
